# [improvement](query) Improve fe high concurrent query performance