@caiconghui 2022-01-30
optimized statistic, 5k connection query 'select 1'
before 561ms 262ms 15.56k
after    406ms 130ms 32.2k